requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.354
langchain-openai==0.0.2
openai==1.3.0
python-multipart==0.0.6
streamlit==1.28.0
faiss-cpu==1.7.4
pypdf2==3.0.1
python-docx==0.8.11
pandas==2.1.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
